pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dockerfile-doctor"
version = "0.1.0"
description = "Lint, analyze, and auto-fix Dockerfiles for best practices, security, and performance"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Alex", email = "islicy@outlook.com" },
]
keywords = ["docker", "dockerfile", "linter", "security", "devops", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
dockerfile-doctor = "dockerfile_doctor.cli:main"

[project.urls]
Homepage = "https://github.com/crabsatellite/dockerfile-doctor"
Repository = "https://github.com/crabsatellite/dockerfile-doctor"
Issues = "https://github.com/crabsatellite/dockerfile-doctor/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
